change tensorrt requirement
