+++ setup.cfg
@@ -38,13 +38,15 @@ install_requires =
     statsmodels~=0.13.5
     tabulate~=0.8.10
     tensorflow==2.13.0
-    tensorrt==8.6.1
     tqdm~=4.65.0
 
 [options.extras_require]
 dev =
     black==22.3.0
     pytest==7.1.2
+
+gpu =
+    tensorrt==8.6.1
 
 [options.packages.find]
 where = src
