+++ src/baskerville/helpers/tensorrt_helpers.py
@@ -1,12 +1,11 @@
 from tensorflow.python.compiler.tensorrt import trt_convert as tf_trt
-from tensorflow.python.saved_model import tag_constants
 import tensorflow as tf
 import tensorrt as trt
 import argparse
 import json
 import numpy as np
 import pandas as pd
-from baskerville import seqnn, layers, dataset
+from baskerville import seqnn, dataset
 
 
 precision_dict = {
@@ -16,62 +15,6 @@ precision_dict = {
 }
 
 # For TF-TRT:
-
-
-class OptimizedModel:
-    """
-    Class of model optimized with tensorrt
-    Args:
-      saved_model_dir: Folder with saved model
-    """
-
-    def __init__(self, saved_model_dir=None, strand_pair=[]):
-        self.loaded_model_fn = None
-        self.strand_pair = strand_pair
-        if not saved_model_dir is None:
-            self.load_model(saved_model_dir)
-
-    def predict(self, input_data):
-        if self.loaded_model_fn is None:
-            raise (Exception("Haven't loaded a model"))
-        # x = tf.constant(input_data.astype("float32"))
-        x = tf.cast(input_data, tf.float32)
-        labeling = self.loaded_model_fn(x)
-        try:
-            preds = labeling["predictions"].numpy()
-        except:
-            try:
-                preds = labeling["probs"].numpy()
-            except:
-                try:
-                    preds = labeling[next(iter(labeling.keys()))]
-                except:
-                    raise (
-                        Exception("Failed to get predictions from saved model object")
-                    )
-        return preds
-
-    def load_model(self, saved_model_dir):
-        saved_model_loaded = tf.saved_model.load(
-            saved_model_dir, tags=[tag_constants.SERVING]
-        )
-        wrapper_fp32 = saved_model_loaded.signatures["serving_default"]
-        self.loaded_model_fn = wrapper_fp32
-
-    def __call__(self, input_data):
-        # need to do the prediction for ensemble model here
-        x = tf.cast(input_data, tf.float32)
-        sequences_rev = layers.EnsembleReverseComplement()([x])
-        if len(self.strand_pair) == 0:
-            strand_pair = None
-        else:
-            strand_pair = self.strand_pair[0]
-        preds = [
-            layers.SwitchReverse(strand_pair)([self.predict(seq), rp])
-            for (seq, rp) in sequences_rev
-        ]
-        preds_avg = tf.keras.layers.Average()(preds)
-        return preds_avg
 
 
 class ModelOptimizer:
@@ -123,7 +66,7 @@ class ModelOptimizer:
 
         converter.save(output_saved_model_dir=output_saved_model_dir)
 
-        return OptimizedModel(output_saved_model_dir)
+        return output_saved_model_dir
 
     def predict(self, input_data):
         if self.loaded_model is None:
@@ -174,7 +117,6 @@ def main():
         strand_transform = dataset.make_strand_transform(targets_df, targets_strand_df)
     else:
         targets_strand_df = targets_df
-        strand_transform = None
     # Load keras model into seqnn class
     seqnn_model = seqnn.SeqNN(params_model)
     seqnn_model.restore(args.model_fn)
@@ -186,7 +128,7 @@ def main():
 
     # Convert the model
     opt_model = ModelOptimizer(f"{args.output_dir}/original_model")
-    model_fp32 = opt_model.convert(f"{args.output_dir}/model_FP32", precision="FP32")
+    opt_model.convert(f"{args.output_dir}/model_FP32", precision="FP32")
 
 
 if __name__ == "__main__":
